One way of specifying filters.  An ugly way.
I thought it'd be wiser to just make use of an existing simple language
and parser, but unfortunately the most promising seemed to be old-style
property lists, but dealing with them is just a pain.

Committing this for posterity in case I ever want to be reminded how
obnoxious this was, before I go off shortly and rewrite it to be way
simpler.